+++ src/InvestmentHub/Providers/Rico/InvestmentHub.Providers.Models.Rico/Responses/GetSummaryPositionResponse.cs
@@ -4,8 +4,8 @@ namespace InvestmentHub.Providers.Models.Rico.Responses
 {
     public class GetSummaryPositionResponse
     {
-        public float TotalValue { get; set; }
-        public float TotalInvestedValue { get; set; }
+        public double TotalValue { get; set; }
+        public double TotalInvestedValue { get; set; }
         public IEnumerable<Position> Positions { get; set; }
     }
 }
